Update README.md
Now a days Government is earning a lot of revenue through the Alcohol. So, the banning of alcohol is impractical and the consumption of Alcohol is increasing nearly 70% for every year. Due to earn more profit sometimes they are not maintaining the standards this is the reason we are seeing a large range of Mental and behavioral disorders. By using this logistic regression we can accurately verify the standards of Alcohol.